Docker/univ-athlete-db/requirements.txt
requests
beautifulsoup4
sqlalchemy
pandas
pytest
discord.py
gspread
oauth2client